requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
psycopg2-binary==2.9.5
pydantic<2.0.0
